# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 智能墨水屏平台 API
  version: 1.0.0
  description: |
    用户端/管理端/设备端统一接口文档。
    设备（真机与模拟）推荐统一启动链路：login -> auto-register -> bind/status -> login -> stream。
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Todo:
      type: object
      properties:
        id: { type: string }
        deviceId: { type: string }
        content: { type: string }
        done: { type: boolean }
        priority: { type: integer, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
...
paths:
...
  /api/todos:
    get:
      summary: TODO列表
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: deviceId
          schema: { type: string }
        - in: query
          name: done
          schema: { type: string }
        - in: query
          name: orderBy
          schema: { type: string }
        - in: query
          name: order
          schema: { type: string }
      responses:
        '200':
          description: ok
    post:
      summary: 新增TODO
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Todo' }
      responses:
        '200':
          description: ok
  /api/todos/{todoId}:
    post:
      summary: 更新TODO
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: todoId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: ok
  /api/todos/{todoId}/delete:
    post:
      summary: 删除TODO
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: todoId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
  /api/todos/batch-upsert:
    post:
      summary: TODO批量保存
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId: { type: string }
                rows: { type: array, items: { $ref: '#/components/schemas/Todo' } }
                deletedIds: { type: array, items: { type: string } }
      responses:
        '200':
          description: ok
  /api/todos/batch-delete:
    post:
      summary: TODO批量删除
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids: { type: array, items: { type: string } }
      responses:
        '200':
          description: ok
...